pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "eqvio"
version = "0.1.0"
description = "Equivariant Visual-Inertial Odometry with Planar Landmarks"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "matplotlib>=3.5",
    "opencv-python>=4.5",
    "liepp",
    "gift",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "sympy",
    "matplotlib",
]

[tool.setuptools.packages.find]
include = ["eqvio*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
